introduces a working modude
reload-cluster now reloads a  appropriate number of processes after a crash an does not run away with many new processes.   There are cases where you can make it kill off newely created proceses. But, this code now prevents that.  It has been hand tested and now the application module needs to be brought back in.