brought start agent back and more fixes to end processes
After disconnect on a process that dies slowly, the cluster.workers will elminate a reference to the worker. But, it is still up to the master to kill the worker if it does not leave on its own.  It is best to request the worker process to be killed as soone as possible.