give attention to logging and tests: 138437710
This is included in test-stress (tests that attempt to
crash the gateway.)

Add in tests to check the performance of the old reload-cluster module.

merge conflicts

updates tests for reload-cluster.js

138437710

Found a bug in the ExitCounter class. Threshold was set, but an undefined
variable was mistakenly used in its place when testing averages. (FIXED)

updates tests for reload-cluster.js

138437710

change a log level in token.js to error.

Found a bug in the ExitCounter class. Threshold was set, but an undefined
variable was mistakenly used in its place when testing averages. (FIXED)

added process id to log message

138852755 Remove unwanted output in Travis tests

138852755 Rename test directory to more meaningful name

give attention to logging and tests: 138437710

parent 1baf91d695aae8c8041600c77515e2daf0f649be
author Ricahrd Leddy <leddyr@google.com> 1564618071 -0700
committer Richard Leddy <richardaleddy@gmail.com> 1565815334 -0700

give attention to logging and tests: 138437710

This is included in test-stress (tests that attempt to
crash the gateway.)

Add in tests to check the performance of the old reload-cluster module.

merge conflicts

updates tests for reload-cluster.js

138437710

Found a bug in the ExitCounter class. Threshold was set, but an undefined
variable was mistakenly used in its place when testing averages. (FIXED)

updates tests for reload-cluster.js

138437710

change a log level in token.js to error.

Found a bug in the ExitCounter class. Threshold was set, but an undefined
variable was mistakenly used in its place when testing averages. (FIXED)

added process id to log message

138852755 Remove unwanted output in Travis tests

138852755 Rename test directory to more meaningful name

139064616 Console log standardization

Pass the component name to core function writeConsoleLog.
parent 1baf91d695aae8c8041600c77515e2daf0f649be
author Ricahrd Leddy <leddyr@google.com> 1564618071 -0700
committer Richard Leddy <richardaleddy@gmail.com> 1565815334 -0700

give attention to logging and tests: 138437710

This is included in test-stress (tests that attempt to
crash the gateway.)

Add in tests to check the performance of the old reload-cluster module.

merge conflicts

updates tests for reload-cluster.js

138437710

Found a bug in the ExitCounter class. Threshold was set, but an undefined
variable was mistakenly used in its place when testing averages. (FIXED)

updates tests for reload-cluster.js

138437710

change a log level in token.js to error.

Found a bug in the ExitCounter class. Threshold was set, but an undefined
variable was mistakenly used in its place when testing averages. (FIXED)

added process id to log message

138852755 Remove unwanted output in Travis tests

138852755 Rename test directory to more meaningful name

139064616 Console log standardization

Pass the component name to core function writeConsoleLog.